src/brain/pyproject.toml
[tool.poetry]
name = "brain"
version = "0.1.0"
description = "The Mind of Voltaire (Python)"
authors = ["Antigravity <ai@voltaire.com>"]

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.50.0"
protobuf = "^4.21.0"
rich = "^13.0.0"
pandas = "^2.2.0"
numpy = "^1.24.0"
alpaca-py = "^0.14.0"
questdb = "^1.0.0"
grpcio-tools = "^1.50.0"
pandas-datareader = "^0.10.0"
requests = "^2.28.0"
torch = "^2.2.0"
# "autogluon.timeseries" = "*"
chronos-forecasting = {git = "https://github.com/amazon-science/chronos-forecasting.git"}
huggingface_hub = "^0.23.0"
feedparser = "^6.0.10"
onnxruntime = "^1.16.0"
transformers = "^4.41.0"
python-dotenv = "^1.0.0"
optimum = {extras = ["onnxruntime"], version = "^1.16.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
